better fixup_utprim and nogood modes

+++ koral.mdefs.h
@@ -33,5 +33,8 @@
 
 
 // precision performance controlled power // ok to use powf with doubles if precision of tolerance is not low.  E.g., ok if tol=1E-6 and kinda ok if tol=1E-9
-#define prpow(x,y) powf(x,y)
-#define prlog(x) logf(x)
+//#define prpow(x,y) powf(x,y)
+//#define prlog(x) logf(x)
+// doesn't help so much, so avoid loss of precision
+#define prpow(x,y) pow(x,y)
+#define prlog(x) log(x)
